Derive the libuast os/arch selector from sys.platform.
Part of the pyuast build process fetches a binary release of libuast.
Previously, the release version was pinned to linux-amd64.

This change makes it consult sys.platform.  Unfortunately, sys.platform doesn't
exactly match the release tag for windows, but since we only have binary
releases for Linux, Darwin, and Windows anyway, it is a straightforward patch.

The default remains linux-amd64 in case someone has another platform; if it was
wrong before it will still be wrong, but in the same way.

Signed-off-by: M. J. Fromberger <michael.j.fromberger@gmail.com>